+++ src/mimic/core/state/src/subnet_index.rs
@@ -46,7 +46,7 @@ impl SubnetIndexManager {
 
     // try_get_canister
     pub fn try_get_canister(path: &str) -> Result<Principal, Error> {
-        let canister = Self::get_canister(path).ok_or(Error::CanisterTypeNotFound {
+        let canister = Self::get_canister(path).ok_or_else(|| Error::CanisterTypeNotFound {
             path: path.to_string(),
         })?;